requirements.txt
wandb
python-dotenv
tqdm
transformers
datasets
numpy
requests